Conflicts:
	Manual/Manual/src/Ejemplo_Entornos_Graficos/Ejemplo_Lamina_Texto.java

+++ src/arquivos/Arquivos.java
@@ -25,17 +25,63 @@ public class Arquivos {
      */
     public static void main(String[] args) {
 
-        // Cadea a1 = new Cadea();
-        Cadea a2 = new Cadea();
+        //  Cadea ejercicio1 = new Cadea(); 
+        // Cadea ejercicio2 = new Cadea();
+        //Cadea ejercicio3 = new Cadea();
+        //Cadea ejercicio3b = new Cadea();
+        // Cadea ejercicio4 = new Cadea(); 
+        // Cadea ejercicio5a = new Cadea(); 
+        //Cadea ejercicio6 = new Cadea(); 
+        //  Cadea ejercicio7 = new Cadea(); 
+        // Cadea ejercicio8 = new Cadea(); 
+        //Cadea ejercicio9a = new Cadea();
+        //Cadea ejercicio9b = new Cadea();
+         // Cadea ejercicio10 = new Cadea(); 
+         
+         
+         // PARA CADA EJERCICIO CONSTRUI UN OBJETO, LA RUTA LA PIDE SIEMPRE POR CONSOLA, SI ACTIBAS TOODS LOS OBJETEOS TE PIDE LAS RUTA DE LOS 10
+         
+         
+                                        //CONSEJO     PROBAR DE UNO EN UNO
+         
 
-        // a1.eDirectorio_eFicheiro();
-        // a2.creaDirectorio();
-        // a2.modoAcceso();
-        // a2.comprobarLonxitude();
-        // a2.mEscritura();
-        // a2.mLectura();
-        // a2.borrarDirectorio();
-        a2.mContido();
+        /*   
+        ejercicio1.creaDirectorio();                                            //F:\Users\david\Documents\NetBeansProjects\arquivos\arquivosdir
+        ejercicio1.eDirectorio_eFicheiro();
+         */
+ /* 
+        ejercicio2.creafichero();                                            //F:\Users\david\Documents\NetBeansProjects\arquivos\arquivosdir\Products1.txt
+        ejercicio2.eDirectorio_eFicheiro();
+         */
+ /*
+        ejercicio3.creaDirectorio();                                                //F:\Users\david\Documents\NetBeansProjects\arquivos\arquivosdir\subdir   
+        ejercicio3b.creafichero();                                                   //F:\Users\david\Documents\NetBeansProjects\arquivos\arquivosdir\subdir\Products2.txt
+         */
+ /* 
+        ejercicio4.mContido();                                                  //F:\Users\david\Documents\NetBeansProjects\arquivos\arquivosdir
+         */
+ /* 
+        ejercicio5a.modoAcceso();
+        ejercicio5a.comprobarLonxitude();                                        //F:\Users\david\Documents\NetBeansProjects\arquivos\arquivosdir\Products1.txt
+         */
+ /*
+        ejercicio6.mEscritura();                                                  //F:\Users\david\Documents\NetBeansProjects\arquivos\arquivosdir\Products1.txt
+         */
+ /*
+        ejercicio7.mEscritura();                                                  //F:\Users\david\Documents\NetBeansProjects\arquivos\arquivosdir\Products1.txt
+         */
+ /*
+        ejercicio8.borrarFichero();                                                 //F:\Users\david\Documents\NetBeansProjects\arquivos\arquivosdir\Products1.txt
+         */
+        /*
+          ejercicio9a.borrarDirectorio();                                           //F:\Users\david\Documents\NetBeansProjects\arquivos\arquivosdir\subdir
+          ejercicio9b.borrarFichero();                                                     //  F:\Users\david\Documents\NetBeansProjects\arquivos\arquivosdir\subdir\texto2.txt
+          ejercicio9a.borrarDirectorio();                                           //F:\Users\david\Documents\NetBeansProjects\arquivos\arquivosdir\subdir
+          ejercicio9a.borrarDirectorio();                                           //F:\Users\david\Documents\NetBeansProjects\arquivos\arquivosdir
+          ejercicio9a.borrarDirectorio();                                           //F:\Users\david\Documents\NetBeansProjects\arquivos
+          */
+        
+        //ejercicio10.recur();                                            //F:\Users\david\Documents\NetBeansProjects\arquivos\arquivosdir
     }
 }
 
@@ -98,6 +144,8 @@ class Cadea {
         } else {
 
             ruta.mkdir();
+
+            System.out.println("Directorio creado en la ruta: ");
             System.out.println(ruta.getAbsolutePath());
 
         }
@@ -108,7 +156,11 @@ class Cadea {
 
         File ruta = new File(r);
 
-        if (ruta.exists()) {
+        if (ruta.isFile()) {
+
+            System.out.println("El archivo ya existe");
+
+        } else {
 
             try {
                 ruta.createNewFile();
@@ -116,9 +168,15 @@ class Cadea {
                 e.printStackTrace();
             }
 
-        } else {
+            if (ruta.exists()) {
 
-            System.out.println("Imposible crear el archivo ruta incorrecta");
+                System.out.println("Archivo creado correctamente");
+
+            } else {
+
+                System.out.println("Archivo no credo");
+
+            }
 
         }
 
@@ -185,10 +243,18 @@ class Cadea {
     public void mEscritura() {
 
         File ruta = new File(r);
+        if (ruta.canWrite()) {
 
-        ruta.setWritable(false);
+            ruta.setWritable(false);
 
-        System.out.println("permisos escritura = " + ruta.canWrite());
+            System.out.println("permisos escritura cambiado = A no permitido");
+        } else {
+
+            ruta.setWritable(true);
+
+            System.out.println("permisos escritura cambiado = A permitido");
+
+        }
 
     }
 
@@ -198,10 +264,23 @@ class Cadea {
 
         if (ruta.exists()) {
 
-            ruta.delete();
+            if (ruta.isFile()) {
+
+                ruta.delete();
+
+                System.out.println("Archivo borrado");
+                System.out.println(" Ruta antigua = " + ruta.getAbsoluteFile());
+
+            } else {
+
+                System.out.println("No se puede borrar es un directorio");
+
+            }
+            System.out.println(" Ruta modificada = " + ruta.getParent());
+
         } else {
 
-            System.out.println("El archivo no existe o ruta incorrecta");
+            System.out.println("Ruta no existe ");
 
         }
     }
@@ -212,13 +291,41 @@ class Cadea {
 
         if (ruta.exists()) {
 
-            ruta.delete();
+            if (ruta.isFile()) {
+
+                System.out.println("No se puede borrar, es un archivo");
+
+            } else {
+
+                ruta.delete();
+
+                if (ruta.exists()) {
+
+                    System.out.println("No se puede borrar directorio con descendencia");
+
+                    File[] listaCarpetas1 = ruta.listFiles();
+
+                    System.out.println(" numero de subcarpetas = " + listaCarpetas1.length);
+
+                    for (int i = 0; i < listaCarpetas1.length; i++) {
+
+                        System.out.println(" rutas de las subcarpetas " + listaCarpetas1[i]);
+                    }
+
+                } else if (ruta.exists() == false) {
+
+                    System.out.println("directorio eliminado");
+
+                }
+
+            }
 
         } else {
 
-            System.out.println("ruta inexistente ou con descencencia");
+            System.out.println("Ruta no existe");
 
         }
+
     }
 
     public void mContido() {
@@ -233,32 +340,45 @@ class Cadea {
 
                 System.out.println(nombre_Contenido[i]); ////muestra el contenido de la ruta dada, tanto directorios como archivos
 
+            }
+        }
+    }
+
+    public void recur() {
+
+        File ruta = new File(r);
+
+        String[] nombre_Contenido = ruta.list();
+
+        if (ruta.exists()) {
+
+            for (int i = 0; i < nombre_Contenido.length; i++) {
+
+                System.out.println(nombre_Contenido[i]); ////muestra el contenido de la ruta dada, tanto directorios como archivos
+
                 File subruta = new File(ruta.getAbsolutePath(), nombre_Contenido[i]); // almacena en el objeto  cada archivo o directorio de la ruta principal
 
                 if (subruta.isDirectory()) { // sirve para saber si la subruta tiene directorios, si es asi  lo almacena en el siguiente array
-                    
-                
 
                     String[] archivos_Subcarpetas = subruta.list(); // almacena en el array los subdirectorios y archivos de la subruta
 
                     for (int j = 0; j < archivos_Subcarpetas.length; j++) {
 
                         System.out.println(archivos_Subcarpetas[j]);  // muestra los archivos y directorios de los subdirectorios
-                        
-                        File sub_sub_ruta = new File(subruta.getAbsolutePath(),archivos_Subcarpetas[j]);  // almacena en el objeto la nueva ruta del subdirectorio encontrado en la ruta anterior
-                        
-                        if (sub_sub_ruta.isDirectory()){
-                            
-                            String[] archivos_sub_sub_carpetas=sub_sub_ruta.list(); // almacena en el array los directorios y archivos de sub subruta
-                            
-                            for (int a=0;a<archivos_sub_sub_carpetas.length;a++){
-                                
-                                System.out.println(archivos_sub_sub_carpetas[a]);  // muestra los archivos y directorios encontrado en  las sub subdirectorios
-                                
+
+                        File sub_sub_ruta = new File(subruta.getAbsolutePath(), archivos_Subcarpetas[j]);  // almacena en el objeto la nueva ruta del subdirectorio encontrado en la ruta anterior
+
+                        if (sub_sub_ruta.isDirectory()) {
+
+                            String[] archivos_sub_sub_carpetas = sub_sub_ruta.list(); // almacena en el array los directorios y archivos de sub subruta
+
+                            for (int a = 0; a < archivos_sub_sub_carpetas.length; a++) {
+
+                                System.out.println(archivos_sub_sub_carpetas[a]);  // muestra los archivos y directorios encontrado en  las sub su
+
                             }
-                            
+
                         }
-                        
 
                     }
 
@@ -267,7 +387,5 @@ class Cadea {
 
         }
     }
-
-  
 
 }
